updates the eos_prefix_list provider
This commit updates the eos_prefix_list provider to fix two bugs found in
smoke testing.  The first issue is the lib code would not return all instances
of ip prefix-list found in the running config.  The second issue was the return
value from lib included extra keys.

Both issues are adreessed in this fix and all prefix-list tests are passing