fixes an issue where the lacp_mode was reset when adding a member interface
This commit fixes a problem with adding a member interface to an exisiting
port-channel and the port-channel interface is configured for lacp_mode value
other than the default.  The lacp_mode would be reset back to the default
mode of on.  After this patch is applied, the portchannel module will now
check the lacp_mode value and apply it when inserting a new member interface
into an existing port-channel